backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
pydantic-settings==2.11.0
openai==1.109.1
qdrant-client>=1.10.1,<2.0.0
python-dotenv==1.0.0
httpx==0.25.2

# Data ingestion dependencies
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.11
pypdf>=3.17.0
docx2txt>=0.8
beautifulsoup4>=4.12.0

